addons/mixed-traffic/d/requirements.txt
fastapi
uvicorn
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi